Implementada teste para a funcionalidade do método patch da entidade usuário
